+++ include/platform/hal_queue.h
@@ -53,14 +53,15 @@ void hal_queue_push_to_back(QueueIdentifier_t queue, const void *item);
  * @param queue The identifier of the Queue that the element should be
  *              inserted
  * @param item  The target item
- * @param timeout After which time (in milliseconds) the receiving attempt
+ * @param timeout After which time (in milliseconds) the "push attempt"
  *		  should be aborted.
- *		  If this value is -1, receiving will block indefinitely
+ *		  If this value is -1 or larger than 9223372036854, pushing
+ *		  will block indefinitely (see hal_semaphore_try_take)
  * @return Whether the attachment attempt was successful
  */
 enum ud3tn_result hal_queue_try_push_to_back(QueueIdentifier_t queue,
 					     const void *item,
-					     int timeout);
+					     int64_t timeout);
 
 /**
  * @brief hal_queue_override_to_back Attach a given item to the back of the
@@ -84,12 +85,13 @@ enum ud3tn_result hal_queue_override_to_back(QueueIdentifier_t queue,
  *		       be stored
  * @param timeout After which time (in milliseconds) the receiving attempt
  *		  should be aborted.
- *		  If this value is -1, receiving will block indefinitely
+ *		  If this value is -1 or larger than 9223372036854, receiving
+ *		  will block indefinitely (see hal_semaphore_try_take)
  * @return Whether the receiving was successful
  */
 enum ud3tn_result hal_queue_receive(QueueIdentifier_t queue,
 				    void *targetBuffer,
-				    int timeout);
+				    int64_t timeout);
 
 /**
  * @brief hal_queue_reset Reset (i.e. empty) the specific queue
